Fix(iad): Read request_type before peeking for WebM header
Modified output_server thread to first read the integer request_type
sent by the client after connecting. Then, it performs a blocking peek
on the socket to check the header of the actual data stream for the
EBML signature. This fixes the issue where the request_type was being
mistakenly checked as the stream header.